requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
lightgbm
catboost
joblib
numpy
python-dotenv
prophetpython